+++ src/main/java/Application/Repos/MoistMeterRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MoistMeterRepo extends JpaRepository<MoistMeter, Long> {
+public interface MoistMeterRepo extends JpaRepository<MoistMeter, Integer> {
 }
