+++ src/main/java/Application/Repos/VideoGameSalesRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VideoGameSalesRepo extends JpaRepository<VideoGameSales, Long> {
+public interface VideoGameSalesRepo extends JpaRepository<VideoGameSales, Integer> {
 }
